Fixed type cast

+++ src/cpp/shared/scripting/funcs.cpp
@@ -823,7 +823,7 @@ static int l_getTownFaction(lua_State *L) {
 
 static int l_setTownFaction(lua_State *L) {
   town* twn = (town*)GetPointerFromLuaClassTable(L, StackIndexOfArg(1, 2));
-  FACTION faction = (FACTION)luaL_checknumber(L, 2);
+  FACTION faction = (FACTION)(int)luaL_checknumber(L, 2);
   twn->SetFaction(faction);
   return 0;
 }
